+++ eos_tax/__init__.py
@@ -1,5 +1,5 @@
 default_app_config = "eos_tax.apps.EosTaxConfig"
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
 VERSION = __version__